+++ tests/test_init.py
@@ -1082,10 +1082,18 @@ class TestOctoRelayPlugin(unittest.TestCase):
         abort_mock.assert_called_with(400, description="Unknown command")
 
     def test_process_at_command(self):
-        # Should toggle the relay having index supplied as a parameter
+        # Should toggle or switch the relay having index and optional target supplied as a parameters
         self.plugin_instance.toggle_relay = Mock()
-        self.assertIsNone(self.plugin_instance.process_at_command(None, None, "OCTORELAY", "r4"))
-        self.plugin_instance.toggle_relay.assert_called_with("r4")
+        cases = [
+            { "params": "r4", "expected": ["r4", None] },
+            { "params": "r4 ON", "expected": ["r4", True] },
+            { "params": " r4  off  ", "expected": ["r4", False] },
+            { "params": " r4  ...  ", "expected": ["r4", None] }
+        ]
+        for case in cases:
+            self.plugin_instance.toggle_relay.reset_mock()
+            self.assertIsNone(self.plugin_instance.process_at_command(None, None, "OCTORELAY", case["params"]))
+            self.plugin_instance.toggle_relay.assert_called_with(*case["expected"])
 
     def test_get_additional_permissions(self):
         # Should return the list of the plugin custom permissions